Divided into modules. Added algorithm and model classes.
